+++ UI/portfolio/src/Components/ProjectsComponent/ProjectsCard.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import GitHubIcon from "../Icons/GitHubIcon";
+import GitHubIcon from "../../Icons/GitHubIcon";
 
 const ProjectsCardDiv = styled.div`
